Fixed rendering performance bug
there was a bug in the calculation of the chunks that should be
rendered, causing the number to increase squared the amount you moved
from origin, leading to really bad performance, when moving the map
further out.

also removed more redundant code. there is a bug when zooming. the
redraw happens based on the origin of the map, making it look like
chunks are calculated wrong, but in reality only the map is moved when
scrolling.